+++ schematic_db/api_utils/__init__.py
@@ -2,5 +2,4 @@
 The Schema class interacts with the Schematic API to create a DBConfig object or to get a list of
  manifests for the schema.
 """
-from .schema import *
-from .database_config import *
+from .api_utils import *
